Move WebsocketManager to websocket package
